Moved proposal mixins to proposals/mixins.py
