+++ proposals/views/proposal_views.py
@@ -25,6 +25,7 @@ from ..forms import ProposalConfirmationForm, ProposalCopyForm, \
     ProposalSubmitForm, RevisionProposalCopyForm, AmendmentProposalCopyForm
 from ..models import Proposal, Wmo
 from ..utils import generate_pdf, generate_ref_number
+from proposals.mixins import ProposalMixin, ProposalContextMixin
 
 
 ############
@@ -236,26 +237,6 @@ onderzoeker of eindverantwoordelijke bij betrokken bent.')
 ##########################
 # CRUD actions on Proposal
 ##########################
-class ProposalMixin(UserFormKwargsMixin):
-    model = Proposal
-    form_class = ProposalForm
-    success_message = _('Studie %(title)s bewerkt')
-
-    def get_next_url(self):
-        """If the Proposal has a Wmo model attached, go to update, else, go to create"""
-        proposal = self.object
-        if hasattr(proposal, 'wmo'):
-            return reverse('proposals:wmo_update', args=(proposal.pk,))
-        else:
-            return reverse('proposals:wmo_create', args=(proposal.pk,))
-
-class ProposalContextMixin:
-        
-    def get_context_data(self, **kwargs):
-        context = super(ProposalContextMixin, self).get_context_data(**kwargs)
-        context['is_supervisor'] = self.object.supervisor == self.request.user
-        context['is_practice'] = self.object.is_practice()
-        return context
 
 class ProposalCreate(ProposalMixin, AllowErrorsOnBackbuttonMixin, CreateView):
     def get_initial(self):
